+++ crates/miden-objects/src/errors.rs
@@ -243,6 +243,8 @@ pub enum AddressError {
     CustomTagLengthNotAllowedForNetworkAccounts(u8),
     #[error("tag length {0} is too large, must be less than or equal to {max}", max = crate::note::NoteTag::MAX_LOCAL_TAG_LENGTH)]
     TagLengthTooLarge(u8),
+    #[error("unknown address interface `{0}`")]
+    UnknownAddressInterface(u16),
     #[error("failed to decode account ID")]
     AccountIdDecodeError(#[source] AccountIdError),
     #[error("failed to decode bech32 string into an address")]